# Compiler support for optimizing PersistentDict