@vchuravy 2023-11-02
this is really cool. A little bit scary but I prefer it over moving the definition into the compiler wholesale.