@topolarity 2023-11-07
What are the consequences of actions that break interface assumptions?

```julia
julia> dict = Base.PersistentDict(0=>false)
Base.PersistentDict{Int64, Bool} with 1 entry:
  0 => 0

julia> dict.trie.data[1] = Base.HashArrayMappedTries.Leaf{Int64, Bool}(0, true)
Base.HashArrayMappedTries.Leaf{Int64, Bool}(0, true)

julia> dict
Base.PersistentDict{Int64, Bool} with 1 entry:
  0 => 1
```

Obviously a user shouldn't do this, but would this now trigger UB?