@Keno 2023-11-09
> Obviously a user "shouldn't" do this, but would this now trigger UB?

Yes, any misuse is undefined behavior. There are some basic sanity checks that avoid applying the optimization if you things are obviously wrong, but we can't really say anything else.