@Keno 2023-11-09
I've given this some further thought, and I don't think the callback design I have here is all that good. I'm planning to rejigger this with an optional return for the getter, having it return `Union{Nothing, Tuple{V}}`. That currently has some performance implications, because in general a tuple return with arbitrary values is a bad idea (there's a reason we turn this into a struct all over the compiler), but we can fix that with a better ABI, which I'll put up as a separate PR.