@aviatesk 2023-11-16
> I don't think the callback design I have here is all that good. I'm planning to rejigger this with an optional return for the getter, having it return `Union{Nothing, Tuple{V}}`

What is the main reason for this? Is it because it makes debugging difficult when users are developing code to utilize this interface?
I'm a bit concerned about the performance impact of `@noinline`s that are necessary for SROA to recognize these interface primitives in cases when SROA can't be performed.