@Keno 2023-11-18
> What is the main reason for this?

The callbacks are a bit awkward, since it's not clear when to infer them, if you want to optimize them.