@Keno 2023-11-20
Review addressed, rebased, and updated in line with what I said I would do above. Needs #52193 to be sound, but other than that, I don't see much reason not to try this. The one remaining concern is that because of the `@noinline` the performance is slightly worse for loosely typed ScopedValues, requiring an extra allocation. However, as indicated above, that can definitely be addressed with some codegen ABI improvements. That said, I'm not planning on finishing those up immediately unless it becomes a problem - I think there's bigger fish to fry elsewhere.