@Keno 2023-11-26
Rebased. I think this is basically good to go. I'm gonna do some more experiments with this, but those might go into a follow up PR, depending.