@Keno 2023-11-27
Working well in my downstream testing. It does still not fully optimize ScopedValue, because we do not yet have the ability to remove all the setup code, which will need some effect tweaks, but that's independent of this, so I'll go ahead and merge this.